readMe file+crud update controller

+++ src/Controller/ApiController.php
@@ -12,6 +12,7 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\Serializer\SerializerInterface;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 
 class ApiController extends AbstractController
 {
@@ -30,7 +31,7 @@ class ApiController extends AbstractController
     }
 
   #[Route('/api/mobile/{id}', name: 'api_show_mobile', methods: ['GET'])]
-  public function mobileShow(Mobile $mobile, SerializerInterface $serializer ): JsonResponse
+  public function showMobile(Mobile $mobile, SerializerInterface $serializer ): JsonResponse
   {
     $jsonContent = $serializer->serialize($mobile, 'json');
 
@@ -39,7 +40,7 @@ class ApiController extends AbstractController
   }
 
     #[Route('/api/mobile', name: 'api_create_mobile', methods: ['POST'])]
-    public function create(Request $request, SerializerInterface $serializer, EntityManagerInterface $em) :JsonResponse
+    public function createMobile(Request $request, SerializerInterface $serializer, EntityManagerInterface $em) :JsonResponse
     {
       $mobile = $serializer->deserialize($request->getContent(), Mobile::class, 'json');
       $em->persist($mobile);
@@ -51,9 +52,32 @@ class ApiController extends AbstractController
 
     }
 
+  #[Route('/api/mobile/{id}', name: 'api_update_mobile', methods: ['PUT'])]
+  public function updateMobile(Request $request, SerializerInterface $serializer, Mobile $currentMobile,
+  EntityManagerInterface $em): JsonResponse
+  {
+    $updatedMobile = $serializer->deserialize($request->getContent(), Mobile::class, 'json',
+      [AbstractNormalizer::OBJECT_TO_POPULATE => $currentMobile]);
+
+    $content = $request->toArray();
+
+    $em->persist($updatedMobile);
+    $em->flush();
+
+    return new JsonResponse(null, JsonResponse::HTTP_NO_CONTENT);
+  }
+
+
     #[Route('/api/mobile/{id}', name: 'api_delete_mobile', methods: ['DELETE'])]
     public function deleteMobile(Mobile $mobile, EntityManagerInterface $em): JsonResponse
     {
+      /*  POP UNE ERREUR VOIR AVEC LE PARAMETER CONVERTER
+     if (!$mobile) {
+       throw $this->createNotFoundException(
+         'There are no mobiles with the following id: '.$id
+       );
+     }*/
+
       $em->remove($mobile);
       $em->flush();
       return new JsonResponse(null, Response::HTTP_NO_CONTENT);
